[Upate]Item,Plane 레이어 추가,Item Scriptable Object 추가

+++ Ephemera/Assets/Script/ItemScript/IItemObtainable.cs
@@ -7,5 +7,5 @@ public interface IItemObtainable
     public void ShowPickupUI();
     public void PickUp(PlayerEx owner);
     public void PickDown(PlayerEx owner);
-    public float SellItem();
+    public int SellItem();
 }
